backend/pyproject.toml
[tool.poetry]
name = "trust-me-bro-backend"
version = "1.0.0"
description = "Trust Me Bro - URL Shortener Backend"
authors = ["Oleh Kazaiev <olehkazaiev@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "0.115.6"
uvicorn = "0.34.0"
sqlalchemy = {extras = ["asyncio"], version = "2.0.36"}
asyncpg = "0.30.0"
pydantic = "2.10.3"
python-multipart = "0.0.19"
redis = "5.2.1"
passlib = {extras = ["bcrypt"], version = "1.7.4"}
bcrypt = "4.0.1"
python-jose = {extras = ["cryptography"], version = "3.3.0"}
alembic = "1.14.0"
greenlet = "3.1.1"

[tool.poetry.group.dev.dependencies]
isort = "5.13.2"
black = "24.3.0"
flake8 = "7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | alembic/versions
)/
'''

[tool.isort]
profile = "black"
line_length = 88
skip = ["alembic/versions"]
